Sorting of classes in dataset generation
Due to the fact that DIGITS uses the standard Python's os.listdir() for retrieving the list of subfolders of the folder with the dataset, the classes are usually stored in random order. This patch adds class sorting with the sorting algorithm from http://stackoverflow.com/a/16090640/1214547 which is a "natural sort", like the one used in Windows Explorer.