Added natural sorting to classes
When a new dataset is added to DIGITS, the parser walks through its folder and considers every subfolder as a class in the dataset. The list of subfolders is retrieved via `os.listdir()`, which does not return a sorted list. This is fixed here, with the sorting key borrowed from http://stackoverflow.com/a/16090640.